implemented adding users

+++ src/main/java/web/model/User.java
@@ -23,13 +23,13 @@ public class User {
     public User() {
     }
 
-    public User(String name, String lastName, int salary) {
+    public User(String name, String lastName, long salary) {
         this.name = name;
         this.lastName = lastName;
         this.salary = salary;
     }
 
-    public User(String name, String lastName, int salary, long id) {
+    public User(String name, String lastName, long salary, long id) {
         this(name, lastName, salary);
         this.id = id;
     }
